requirements.txt
# Arabic Sign Language Translator — Dependencies

flask==3.1.3
requests==2.31.0
numpy==2.3.4
sentence-transformers==5.4.1
openai-whisper==20250625
sounddevice==0.5.5
scipy==1.17.0
torch==2.11.0
transformers==5.6.2
